+++ src/widgets/header/ui/header-user-actions/header-user-actions.module.css
@@ -1,10 +1,3 @@
-.user-controls {
-  position: relative;
-  display: flex;
-  gap: 3px;
-  align-items: center;
-}
-
 .button-link {
   padding: 8px 16px;
   border-radius: 10px;
@@ -18,7 +11,15 @@
 }
 
 .button-link:hover {
+  text-decoration: none;
   background-color: rgb(201 196 196);
+}
+
+.user-controls {
+  position: relative;
+  display: flex;
+  gap: 3px;
+  align-items: center;
 }
 
 .profile-icon,
