+++ src/widgets/header/ui/header-user-actions/header-user-actions.tsx
@@ -1,8 +1,8 @@
-import IconArrow from '../../../assets/svg/arrow.svg';
-import IconBasket from '../../../assets/svg/basket.svg';
-import IconProfile from '../../../assets/svg/profile-icon.svg';
+import IconArrow from '../../../../assets/svg/arrow.svg';
+import IconBasket from '../../../../assets/svg/basket.svg';
+import IconProfile from '../../../../assets/svg/profile-icon.svg';
 import { Link } from 'react-router-dom';
-import style from './HeaderUserActions.module.css';
+import style from './Header-user-actions.module.css';
 import { useState } from 'react';
 
 export function HeaderUserActions() {
