+++ src/widgets/header/ui/header/header.tsx
@@ -1,0 +1,24 @@
+import { HeaderUserActions } from '../header-user-actions/Header-user-actions';
+import { Link } from 'react-router-dom';
+import style from './header.module.css';
+
+export const Header = () => {
+  return (
+    <header className={style.header}>
+      <nav className={style.navigation}>
+        <Link to="/" className={style['button-link']}>
+          <h1 className={style.title}>Techhouse</h1>
+        </Link>
+        <div className={style['menu-page']}>
+          <Link to="/catalog" className={style['button-link']}>
+            Catalog
+          </Link>
+          <Link to="/about" className={style['button-link']}>
+            About Us
+          </Link>
+        </div>
+        <HeaderUserActions />
+      </nav>
+    </header>
+  );
+};
