--- src/widgets/headerUserActions/index.ts
@@ -1,1 +1,0 @@
-export { HeaderUserActions } from './ui/HeaderUserActions';
